# docs/swagger.yaml
basePath: /
definitions:
  handler.errorResponce:
    properties:
      message:
        type: string
    type: object
  handler.statusResponse:
    properties:
      status:
        type: string
    type: object
  models.Address:
    properties:
      address:
        type: string
    type: object
...
  models.Status:
    properties:
      status_id:
        type: integer
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API server for online shop application
  title: Online Shop App
paths:
  /api/admin/orders:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
      security:
      - ApiKeyAuth: []
      summary: all orders
      tags:
      - Admin
  /api/admin/orders/{id}:
    put:
      consumes:
      - application/json
      parameters:
      - description: order id
        in: path
        name: id
        required: true
        type: string
      - description: new status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/models.Status'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.statusResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
      security:
      - ApiKeyAuth: []
      summary: edit order status
      tags:
      - Admin
...
  /api/cart/order:
    post:
      consumes:
      - application/json
      parameters:
      - description: address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/models.Address'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
      security:
      - ApiKeyAuth: []
      summary: make order
      tags:
      - order
  /api/cart/orders:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.errorResponce'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handler.errorResponce'
      security:
      - ApiKeyAuth: []
      summary: order history
      tags:
      - order
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"